refactor!: split `mod.rs` and `test_utility.rs` between the two new modules
