+++ crates/proof-of-sql/src/sql/mod.rs
@@ -1,5 +1,6 @@
 //! This module contains the main logic for Proof of SQL.
-pub mod ast;
+pub mod proof_expr;
+pub mod proof_plan;
 pub mod parse;
 pub mod postprocessing;
 pub mod proof;
